+++ src/hittable.rs
@@ -6,7 +6,7 @@ use crate::vec3::{dot, Point3, Vec3};
 pub struct HitRecord {
     pub pt: Point3,
     pub normal: Vec3,
-    pub t: f32,
+    pub t: f64,
     pub is_front_face: bool,
 }
 
@@ -15,7 +15,7 @@ impl HitRecord {
     /// value for the given ray. This method should be called during geometry
     /// computation.
     pub fn set_face_normal(&mut self, r: Ray, outward_normal: Vec3) {
-        assert!(f32::abs(outward_normal.length() - 1.0) < 0.01);
+        assert!(f64::abs(outward_normal.length() - 1.0) < 0.01);
         self.is_front_face = dot(r.direction(), outward_normal) < 0.0;
         self.normal = if self.is_front_face {
             outward_normal
