+++ src/vec3.rs
@@ -1,11 +1,11 @@
-use core::f32;
+use core::f64;
 use std::ops::{Add, Div, Mul, Neg, Sub};
 
 use crate::random::Rng;
 
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub struct Vec3 {
-    e: [f32; 3],
+    e: [f64; 3],
 }
 
 pub type Point3 = Vec3;
@@ -14,7 +14,7 @@ pub fn unit_vector(v: Vec3) -> Vec3 {
     v / v.length()
 }
 
-pub fn dot(u: Vec3, v: Vec3) -> f32 {
+pub fn dot(u: Vec3, v: Vec3) -> f64 {
     u.e[0] * v.e[0] + u.e[1] * v.e[1] + u.e[2] * v.e[2]
 }
 
@@ -25,40 +25,40 @@ impl Default for Vec3 {
 }
 
 impl Vec3 {
-    pub fn new(e0: f32, e1: f32, e2: f32) -> Self {
+    pub fn new(e0: f64, e1: f64, e2: f64) -> Self {
         Self { e: [e0, e1, e2] }
     }
 
     pub fn new_random_unit_vector(rng: &mut Rng) -> Self {
-        let pi = f32::consts::PI;
+        let pi = f64::consts::PI;
         let phi = pi * rng.generate();
         let theta = 2.0 * pi * rng.generate();
 
-        let x = f32::sin(phi) * f32::cos(theta);
-        let y = f32::sin(phi) * f32::sin(theta);
-        let z = f32::cos(phi);
-        assert!(f32::abs(x * x + y * y + z * z - 1.0) <= 0.001);
+        let x = f64::sin(phi) * f64::cos(theta);
+        let y = f64::sin(phi) * f64::sin(theta);
+        let z = f64::cos(phi);
+        assert!(f64::abs(x * x + y * y + z * z - 1.0) <= 0.001);
 
         Self { e: [x, y, z] }
     }
 
-    pub fn x(&self) -> f32 {
+    pub fn x(&self) -> f64 {
         self.e[0]
     }
 
-    pub fn y(&self) -> f32 {
+    pub fn y(&self) -> f64 {
         self.e[1]
     }
 
-    pub fn z(&self) -> f32 {
+    pub fn z(&self) -> f64 {
         self.e[2]
     }
 
-    pub fn length_squared(&self) -> f32 {
+    pub fn length_squared(&self) -> f64 {
         self.e[0] * self.e[0] + self.e[1] * self.e[1] + self.e[2] * self.e[2]
     }
 
-    pub fn length(&self) -> f32 {
+    pub fn length(&self) -> f64 {
         self.length_squared().sqrt()
     }
 }
@@ -89,7 +89,7 @@ impl Sub for Vec3 {
     }
 }
 
-impl Mul<Vec3> for f32 {
+impl Mul<Vec3> for f64 {
     type Output = Vec3;
     fn mul(self, rhs: Vec3) -> Vec3 {
         Vec3 {
@@ -103,17 +103,17 @@ impl Mul<Vec3> for u32 {
     fn mul(self, rhs: Vec3) -> Vec3 {
         Vec3 {
             e: [
-                self as f32 * rhs.e[0],
-                self as f32 * rhs.e[1],
-                self as f32 * rhs.e[2],
+                self as f64 * rhs.e[0],
+                self as f64 * rhs.e[1],
+                self as f64 * rhs.e[2],
             ],
         }
     }
 }
 
-impl Div<f32> for Vec3 {
+impl Div<f64> for Vec3 {
     type Output = Self;
-    fn div(self, rhs: f32) -> Self {
+    fn div(self, rhs: f64) -> Self {
         Vec3 {
             e: [self.e[0] / rhs, self.e[1] / rhs, self.e[2] / rhs],
         }
